requirements.txt
# Python=3.12 is recommended

pandas==2.3.3
torch==2.9.0
torchvision==0.24.0
sentence-transformers==5.1.1
faiss-cpu==1.9.0
tqdm==4.67.1
requests==2.32.5

